# Model Syntax nodes as structs