@ahoppen 2019-10-16
@swift-ci Please test