@ahoppen 2019-10-17
@swift-ci Please test